create ~/Notes folder at start
Warning: I'm still very green with the project; this PR should be extensively reviewed.

Starting `sploitkit-new` fails with an obscure sqlite error. The cause of the error is a missing parent directory in the path, more specifically the 'WORKSPACE' path as defined by default in the options.

I added a check for a missing directory, mirroring the functionality of "app_folder" to "workspace" (which means "workspace" can now be moved).